breaking: add brokerWildcard config
Creating pub-sub worker handler topic paths as filenames with '*' in the
filename is not possible on all filesystems. To enable development on
all filesystems a new config property has been added: brokerWildcard.

It is not required but defaults to '%' (which is usable on all
filesystems in filenames). '*' can still be set as borkerWildcard, but
it is advised to use '%' from now on. When upgrading from 2.x it is
necessary to set the brokerWildcard config property to '*' when worker
handlers with a '*' in the filename are being used.